requirements.txt
requests>=2.31.0
pandas>=2.0.0
openai>=1.3.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
python-dateutil>=2.8.2
weasyprint>=60.1.0  # For PDF generation
